Added canvas
New CanvasLayout can be used to place collection of objects in some
area. It doesn't change positions of objects but only sets parent box
for them. Also it's possible to add, remove and replace objects.
Built by design objects are placed in this layout. They behave as common
objects and can be replaced with new designed objects by clicking on
button "Update".