Added Presentation
All enums and classes should have corresponding Presentation that
describes their values, members, base classes, etc. Presentation is
serializable so it can be serialized by DesignViewBuilder to TreeView
and view of properties. But it's needed to create the first Presentation
that describes enums and classes used in Presentation. This first
Presentation is hardcoded. Presentation of all other types should be
created in design_editor.