TextList's ButtonList is always on top
Added system top view view with maximum zIndex that contains transparent
CanvasLayout. ButtonList is added on this canvas when it's opened. To
save position and box of original TextList new class ObjectReflection is
used. It reflects position and box of another object and can be parent
of another objects as if it's original object.